requirements.txt
anyio==3.5.0
asgiref==3.5.1
cassandra-driver==3.25.0
click==8.1.3
fastapi==0.75.2
geomet==0.2.1.post1
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
pydantic==1.9.0
python-dateutil==2.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.2.0
unicorn==1.0.3
uvicorn==0.17.6
zipp==3.8.0
